First major update for reading the input file
A major change that I made was that the input file now is read in a strict order.

Also changed many comments that were not in proper english. Also changed some variable names that did not reflect the actual value or purpose they had. This involved changing almost all the other files.

Another major change was that I mode the particle type from Sample_Geometry to Simulation_Parameters. It makes more sense to me to have the type of nanoparticle as part of the simulation parameters than the geometry of the sample, although some geometry is being specified. This also involved changing several other files. Some changes are still pending, which will be done in a following update.

Also updated the comments in the input file.