Restored use of string stream
It truns out that setting a string stream using str(string) does not work for getting the values read from the file. The values can only been extracted when the string stream is initialized (at declaration) with the line read from the file. Or at least that is what I observed.

Thus I removed all usages of the str function and used multiple string streams.