Updated matrix density variable and usage
There were two variables for the matrix density: matrix_density in Nanotube_Geo and polymer_density in Geom_sample.

I kept the name matrix_density to keep the idea of a general matrix not exclusively a polymer. However, I kept this variable in Geom_sample as it is pretty much a property of the sample.

Also, the calculation of the generated weight fraction was not accurate as the volume of nanoparticles was ignored when calculating it. I corrected this.