Updated variables for electrical parameters
I changed the name of the variable for CNT resistivity from resistivity_CF to resistivity_CNT. We never used carbon fiber (CF) so now it is more illustrative.

Also removed the sheet resistance of the GNP which was never used. If needed it probably can be calculated from the two resistivities and geometry.