Updated some parameters for GNP_Geo
I added a new option to specify CNT density per unit area. This was done before by specifying a negative mass ratio. Now the mass ratio in the input file always has to be positive. I think it makes more sense to have this new option than havign to specify a negative number as input of a positive value.

Also, now the definition of the functions to read the input file are in the same order as the sections in the input file.