Romoved points outside sample right after generation
Now, the points outside the generation domain are removed right after all points for all CNTs are generated. This will reduce the memory needed in the execution of the program.

A minimum number of points needs to be specified. For now it is an internal variable, but needs to be an input variable.