Corrected and simplified window extraction
I made a new function to extract an observation window. This new function is based in the function used to remove the CNTs outside the sample (in the boundary layer). I found there were some errors in this function (the one that removes CNTs outside the sample), which I corrected.

I also added a function to calculate the geometry of the observation window (in App_Network_3D.cpp). I think this makes that class easier to read.