Updated shells
After removing the CNTs in the boundary layer, I could not use the shell that correspont to that layer. This beacuse only the CNTs in the layer but not in the boundary are included in that shell. So I modified the function to include the CNTs in the boundary of the sample as part of the shell for the boundary layer.

Also, I made some modifications in the functions for adding a CNT to a shells. I precomputed some values and pass them as arguments. This will reduce the number of computations. For small examples this seems negligible. Maybe for larger samples this might be noticeable. Although, most of the computation time is for solving the system of equations. So even for large sample this change might not be noticeable. Still, it is more efficient.