Separated functions for transformation of point vectors
I now have two functions for transforming points. One for CNTs and one for GNPs. This beacuse I needed some variables specific for CNTs when calling the function for GNPs. So it made more sense to make its own function. Then if it needs to be udpdated, I can just add new arguments.

I also found and corrected a mistake. I was using the vector of radii for the generation domain. So I created a new variable to generate a radii vector for the CNTs inside the sample.