Seeds for GNPs
Now I can specify seeds for generating the GNP network. So now I can generate the same GNP network using the seeds for the random generator number. Before, this was only possible for CNTs.

Now there are two vectors for the seeds, one for CNTs and one for GNPs.

I also changed the way the seeds are output. Now they are all in a single line. In this way, to re-run a simulation with the same network I can now just copy paste all seeds at once.

Also updated the input file indicating how many seeds are needed and how to input them.