GNP generation without interpenetration
Most of the functions for GNP generation are finished. I am missing the GJK and EPA. I left comments for the functions that need to be written and need the GJK/EPA output. I did this update in this way beacuase now I need to implement GJK and EPA and I prefer to have that clase on its own update. So the next update will be GJK and EPA and then I wil finish the GNP generation function.

I also made some improvements to the Geometry_3D class. I added the normal to the plane in the Plane_3D class. I also added a function to obtain a string output for a Plane_3D object (just in case I need it).