Corrected direction vector
For determining the distance between polyhedra, I started with a point inside the Minkowski sum. Then what I need is the vector from this internal vector towards the origin. I was using the point as the vector, i.e., I was actually using the vector from the origin towards the internal point. So I corrected this.