Added some debugging comments
I added some debugging comments. There were actually not needed as the problem was in an input, not in the code.

The only comment that is actually printed to file gives information about the number of CNTs and points in the generation domain. In this way the number of CNTs and points in the generation domain and in the composite domain can be compared (CNTs and points in the composite domain are already printed to file).

Also made a small change and removed a variable that was not needed.