Corrected distance between GNPs
After correcting the GJK I also needed to correct the way in which I calculated the distance between GNPs since the same idea was used.

At the end I realized that the simplex in the final iteration (after one last update) can be used to get the distance and separation direction between two non-penetrating GNPs.

Thus, almos all functions for estimating distance were deleted. The one I left is to calcualte the distance and direction from a simplex to the origin. Of course, this simplex is the one given as output from GJK.