Corrections in extracting the observation window for CNTs
The functions that extracted the observation window for CNTs used the same logic that I used for removing the CNTs in the boundary layer. Since I made some corrections on the functions for the boundary layer, I also had to make some corrections for extracting the observation window.

The new function for calculating the point at the boundary was also updated.

Also, I also made some other corrections on how to handle segments. It is more like in the previous version of the code, where two segments where separated by only one outside point. That is, the last point of one segment was outside, and this was the first point of the new segment. In such case I just move the point to the boundary while handling the first segment. Then this point is shared with the next segment.