Updated shells for GNPs
Updated the function that adds GNPs to shells for extracting an observation window. Now this can be done using the GNP data structure. Deleted the old versions.

I also changed some function names so that they are more meaningful. This resulted in changing Cutoff_Wins.cpp.

I added a flag to the GNP to include its number (i.e., its index in the vector of GNPS, which serves as an identifier).