Nanoparticle content as a single input
Now the nanoparticle content is specified in the simulation parameters (Simu_para). Before there was a field for CNTs and one for GNPs. Also, the mass ratio and density of CNTs over GNPs is part of simulation parameters. I added the option for volume ratio. In this way it is more clear what quantities are input and confusion is avoided by now having a single field for content.

The input file was modified accordingly. The content fileds were removed from CNT and GNP parameters. Now they are part of simulation parameters.

Another important change is that
now the generation is done only considering the nanoparticle volume. In this way only one termination criteria is checked instead of two (that actually requires at least three conditions). So, when weight fraction is specified, the volume fraction and volume are calculated. Also, it is easier to work with volumes than weights. Weight and weigth fraction requires density and the weight of matrix. In contrast, using volume fraction does not require the matrix volume, the sample volume is all that is needed. These changes are tested for CNTs only and GNPs only. After updating mixed and hybrid particles more changes could be made.