GNPs in the boundary layer are ignored
Now if a GNP is completely outside the sample (i.e., in the boundary layer), then it is not added to the vector of GNPs. This are the "ignored" GNPs. I added a variable to count these ignored GNPs.

I also improved the way in which the information regarding the number of generated, rejected, and ignored GNPs is presented.