CNTs in the boundary layer are ignored
CNTs that are completely outside the sample are not added to the CNT and point vectors. They are simply ignored. This reduces the memory usage. I also added a variable to count the CNTs that are ignored. Also, the CNTs that don't grow beyond the seed are included in the count of rejected CNTs.

Because of CNTs completely outside the sample are ignored, the count of CNTs in the generation domain is not correct anymore. So, I updated the information about generated CNTs.