Corrected count of rejected CNTs
I had added a condition that when the new CNT vector has 1 or 0 points then the count of rejected CNTs is increased. Actually, when a CNT is rejected the vector cnt_new is cleared and the reject count is increased. Thus, a rejected CNT always results in the new CNT vector having 0 points. Thus, I was counting twice. I removed this condition as it turned out to be redundant.