Completed code for Mixed particle generation
Completed the code for mixed particle generation. I copied the CNT generation as a base and added functions for CNT points interpenetrating GNPs. Of course there is still debugging to do.

I also modified or deleted some old functions.

I also added a function to the Plane_3D class to measure the distance from a point to a plane.

Now there are two overlapping lengths: one for CNTs and one for GNPs. In this way, overlapping regions  of the same size can be used for both.