Sample geometry as a cuboid
Added a cuboid to Geom_sample for the sample domain geometry. In this way I can check if a point is inside of the sample domain or extended domain. I also added maximum coordinates to the cuboid structure and deleted volume as it is actually never used.

The cuboid for the sample is used for nanoparticle generation and for constructing the shells. Shells are updated in Cutoff_Wins, so I made a change there but just for the shells. I will be phasing out the usage of coordintes and origin as I progress with the changes.

The same can be done with a window, instead of using a Geom_sample I can simply use a cuboid. This change will have to come later.