Removed repeated functions from Geometry_3D
Removed some of the functions in Geometry_3D for the Point_3D class that were repeated. The difference in these repeated functions was the usage of const and the & symbol. I just left the one that was actually needed.

I also removed the data structure for ellipsoids as it is not used anymore.