Debugging Cutoff_Wins
The condition for checking if a coordinate was between the bounds of the observation window probably was not correct. I changed it to an expression that is indeed correct.

When checking if the point's z-coordinate was bounded by the sample z-coordinates, I was actually using the point's y-coordinate. Thos was corrected.

I was missing the initialization of the boundary vectors for GNPs.

Added some comments for debugging.