Missing condition in Generate_Network for mixed particles
I was missing a condition in Generate_Network for mixed particles. When a CNT seed is generated, new_cnt is empty. So, if CNT seed is generated inside a GNP it will be moved. After the CNT seed is moved, then there is a check for segment orientation. Since the new_cnt vector is empty, there is a segmentation fault. Thus, I added the condition to check the size of the new_cnt vector.

Added some comments for debugging.