Modified size of CNT overlapping
The variable in Input_Reader for CNT overlapping was set as the one for finding penetrating points during generation. However, I modified so that it takes the one needed for finding contacts. In this way I reduce the number of arguments in the Contact_grid function Generate_contact_grid. Also, in the future, I could use the sectioned domain from generation directly for finding contacts.

I also corrected a typo in a variable name in Input_Reader class, I changed  cnt_gnp_densinty to cnt_gnp_density.