Hoshen-Kopelman Update
I re-wrote the class for the Hoshen-Kopelman algorithm. It now calculates contact points for GNPs since their discretizations are not used anymore.

Other important changes include that consecutive contacts between the same particles are "compressed".  For this, I added the variable min_points to the cutoff distances data structure. This variable determinas the minimum number of points for two contacts to be considered too close. And if the contacts are too close, they are "compressed" into one contact. Also, a CNT point can have only one contact with another CNT. This eliminates the case of a CNT point with two tunnels with two different points in the same CNT.

I also created a new data structure for junctions. This data structure is part of the Geometry_3D class.