Added percoaltion to Hoshen_Kopelman class
Added the function to determine percolation to the class Hoshen_Kopelman since I found that it did not require that much code. Compared to the previous version it does seem like a great simplification of the code. Because of this change the vector of families is in this class now.

I also deleted a function that I had added to the Hoshen_Kopelman class but ultimately did not use.