Corrected a distance calculation
Whe generating GNPs using the non-penetrating model, it may happen that two generated GNPs are touching. In such case, I find the simplices in the GNPs that are touching. To do so, I calculate the distance from the centroid of the Minkowski sum (A+(-B)) to the simplex that contains the origin. I had made an error when calcualting this distance in which I had a wrong sign. This is corrected in this update.

I also added a few comments to explain what is being calculated.