Added a variable for CNT points at boundaries
Added a variable to Cutoff_Winds for the CNT points at each boundary. I already had the same variable for GNP points and was missing the one for CNT points.

As a consequence of using this new variables I deleted two functions that updated the old variables for points at the boundary. I have not deleted the old variable because it will cause error in other classes.