GNP points and structure in Cutoff_Wins
The GNP points at boundaries were stored in a class vector in Cutoff_Wins. However, it will be more helpful if all points are in a single vector. So, in this change the GNP points at boundaries are stored in the vector of GNP points.

Also, I am generating the GNP structure as GNP points are generated. I realized that generating the structure as pre-processing was not optimal. So the GNP structure is filled in Cutoff_Wins and Hoshen_Kopelman as points are generated.