Structure and junctions in Hoshen_Kopelman class
Now the GNP structure is filled in the Hoshen_Kopelman class as GNP points are added to the vector points_gnp.

Also, for compatibility I changed the type of the vectors in structure_gnp to long int. This will make possible to use the same functions as structure CNT, since that vector also has vectors of long ints. As a consequence of this change, I updated the GNP structure in the class Cutoff_Wins. In addition, in Cutoff_Wins, the vector boundary_gnp_pts is now a vector of long int vectros.

Anothe change made here is that there are now three vectors to store the junctions: one for CNT-CNT junctions, one for GNP-GNP junctions, and one for CNT-GNP junctions. I think this will make the code faster as I will not need to check each junction to find out wich type of junction it is.