Updated Triangulation class
Updated the Triangulation class so that it now uses the edges and triangular faces data structures that I created for the GJK algorithm. However, the data structures used for GJK use int to store vertices. For the triangulations I need long int as this is how point numbers are stored. So I created EdgeL and TrFaceL as the same data structurs as Edge and TrFace, respectively, but using long int to store point numbers.