Removed the usage of CNT structure in DEA
I noticed the elements made during the Hoshen-Kopelman did not have the first and last points. Then, during the DEA they were taken directly from the CNT structure (vector structure_cnt).

So, I modified the Hoshen-Kopelman to also add the first and last points to each CNT element that has at least one junction. With this change, the CNT structure is not needed anymore for the DEA.

While removing the CNT structure from the DEA I noticed I was filling the stiffness matrix with CNT resistor incorrectly. I was not advancing the first point of each element formed by two consecutive points in the vector of elements. In that way, all elements were the first point of the CNT and the current point iterated. So I corrected this.