Updated backbone extraction and efficiency calculation
Updated the class Backbone_Network which extracts the backbone.

Now, the valumes of nanoparticles in each percolated family are calculated in this class instead of its own class. The fractions of each percolated family which are used to calcualte efficiency will then calcualted in the Electrical_analysis class in a future update.

I also updated the Junction data structure as it became very clear that I needed it to have the particle number too. So now it contains the two particle numbers.