GNP volume updated
During generation, the approximated GNP volume inside the sample is updated in the variable volume of the GNP data structure.

Also, during window extranction, the GNP volume is updated as it might be partially outside the observation window. To do this, I had to modify the function in Generate_Network that approximated the volume inside a cuboid.