Update in triangulation and their electrical resistance
To save time, triangulation are performed in two cases:
1) When the DEA is used for backbone extraction
2) When de DEA is used for calcualting the electrical resistance but only for the first time if the cluster percolates in more than one direction. That is, if the electrical resistance is calculated and the triangulation has not been cleared.

I also changed the use of the map points_cnt_rad to a simpler one. It used to be a map of the type P_gnp->P_cnt, where P_gnp was a GNP point and P_cnt a CNT point. Then, using the CNT poit I used the vector points_cnt to obtain the CNT number and then with the radii vector I could obtain the CNT radius. This radius is needed to calculate the electrical resistance of the triangulation edge.

However, intead of doing it this way I coul just use the map P_gnp->r_cnt, where r_cnt is the CNT radius. Since this is faster and requires the use of less variables in the function, now this is how the map points_cnt_rad works.