Calculation of fractions
Now fractions of each percolated family is done in the class Electrical_analysis. Only two files are printed, volumes and fractions. When mixed or hybrid particles are used the volume of total nanoparticles are exported.

Also, I added a flag to decide whether CNT and GNP volumes and fractions are also printed into a file when using mixed ot hybrid particles.