Added call to Electrical_analysis in App_Network_3D
Added the call to the class Electrical_analysis in App_Network_3D. I also noticed that the vector family now is part of Hoshen_Kopelman, so it was a redundant input parameter. So I modified the classes Electrical_analysis and Backbone_Network so that the family vector is used from the Hoshen_Kopelman object.

I also removed all the unused functions from Electrical_analysis and the class variable it had.