Deleted deprecated code
Deleted all the functions and class variables that were not used anymore due to the update of the code.

Although this was supposed to be an update where only code is deleted and no changes are made, I realized that I was still using an old function. So the only change besides deleting old code was updating the calls to a function that calculates the dot product of two vectors in the conjugate gradien algorithm.

I also changed the GS to GNP in the header comment of each file.