Visualization files for backbone and dead particles
I added the functions to export visualization files for CNTs and GNPs in the backbones and dead branches (CNTs) and dead particles (GNPs). I also added the functions to export the isolated particles, which are all the CNTs and GNPs that are not part of a percolated cluster.

Thus, when activating the flag for the backbone, up to six files are generated: CNTs in backbone, dead branches and isolated; and GNPs in backbone, dead (attached to the percoalted cluster) and isolated.

I also updated the comments in the input file for the flag that controls the export of the backbone.