Added flag to save CNT and GNP fractions and volumes separately
Added the reading of the flag that saves CNTs and GNPs volumes and fractions to the Input_Reader class.

If only CNTs or only GNPs are generated, this flag is ignored by setting it to zero.

I also added the flag into the input file.

I added some "return 0" statements that where messing when errors while reading the flags.