Clearing structure_cnt[CNT] of a CNT outside the observation window
I was doing nothing to identify a CNT outside the observation window when extracting it. Thus, some CNTs outside the observation window caused a segementation fould when being added to the sectioned domain.

Then, here I clear the vector structure_cnt[CNT] which was already used to determine if a CNT was inside the observation window or not.

I also added some comments for debugging. In some functions I removed some parameters that were not being used.