Corrected update of CNTs trimmed when extracting observation window
When extracting the observation window, I noticed there were two bugs that cuased a segmentation fault in the Hoshen-Kopelman algorithm. While finding the cause of thse bugs I found a third one. These are:

1) When trimming some points at the front of the CNT beacuase they are outside the observation window, the CNTs that remain inside the window are moved to the front. Here, I was copying alwyas the same point. These resulted in having a vector structure_cnt[CNT] where all points had the same point number. This was corrected.

2) When calcualting and substituting a point at the boundary, I was not updating the CNT flag, i.e., the CNt number. This resulted in CNTs with negative flag, i.e., negative CNT number. Thus when using this flag as an index for a vector I had a segmentation fault. This was corrected.

3) I found that when adding a new structure vector I was using the index instead of the point number. This was corrected.
In addition, I realized that I was not calling the function that adds the first and last points of the CNT elements. So I added this call.

Also, I added comments for debugging.