Corrections in compressing junctions
Two bug were causing a semgnetation fault while compressing the CNT-CNT junctions. First, the vector of elements was not initialized to its proper size. Second, I was not using the vector contact_elements properly. I needed to check if it had elements in element i. However, I was always checking if it had elements. These two were corrected.

I also found that I was using a long int instead of double to store the junction distance. This caused the junction distance to always be zero. This was corrected.
Initialize the vector of elements

I also added som comments for debugging.