Update on making the percoalted clusters
I found that I was not adding the CNTs at the boundary on the boundary vector (when extracting an observation window). This resulted in never having percolation. For this I updated the Cutoff_Wins class.

In the Hoshen_Kopelman class I was having a segmentation fault error when adding the junctions to a cluster. This happened because the vector of cluster junctions is as big as the nuber of percolated clusters. So if there was a junction in a non-percoalted cluster, it would be in a cluster that is not in the vector of cluster junctions. I added a map to solve this issue, where a cluster label is mapped to the percolated cluster number.