Forgot to add points to elements
In class Hoshen_Kopelman, when compressing junctions I formed the CNT elements. However, in one part of the code I was adding a compressed junction but not the points into de elements. This caused a segmentaton fault in the electrical analysis when an empty set was being scanned to form the LM matrix.

Thus, I added the lines to insert the junction points into the CNT elements.

I also added some comments for debugging.