Incorrect use of points vector
I was having some segmentation fault errors because, at some point, I was using the vector of GNP points instead of the vector of CNT points whent testing only CNTs. I changed this to using the vector of CNT points.

Also, while adding the CNT-CNT junctions to the stiffness matrix, I found that I was not using the particle numbers from the Junction object. Instead, I was using the flag from the point object. I changed this to using the particle number from the Junction object. I also made this correction for GNP-GNP and CNT-GNP junctions.