Corrected column index in SSS
The vector col_ind (column index) for the SSS format of the sparse stiffness matrix was not being updated with the correct column numbers. I was using the ones for the original stiffness matrix, however, the column indices needed are those of the "reduced" stiffness matrix. To do this, I just subtracted the number of reserved nodes.

I also added some comments for debugging and printed some vectors into a file. I also added some commands to fill the vector of all currents in all elements. However, I am not adding the currents passing through CNT-CNT and GNP-GNP junctions.