Unit resistances for junctions
I was actually not using unit resistors for junctions at all. This because the functions that added the resistors due to junctions did not have the resistance flag as an input.

So I added the resistance as an input to the functions that add the resistors for the junctions. I also added the if-statements that set the unit resistor or calculate the electrical resistance depending on the value of the resistance flag.

I also added some functions to add all currents of all resistors into a vector, sort it and print it into a file.