Flags for calculating the electrical resistance on each direction
Added three flags for calculating the electrical resistance on each direction.

There still is the flag avoid_resistance, which would act as a main flag. Then, three more flags are used, one per each direction x, y, z. For these three new flags, 1 means calculate the resistance on that direction and 0 means not to calculate the resistance in that direction. In this was the resistance can be calculated in only one direction instead of 3 in the cases where there is percolation along the three directions.