Eliminated the calculation of matrix KEFT
In the conjugate gradient, to obtain an initial residual, I generated the matrix KEFT and then multiplied it by the vector of prescribed voltage. Which is what has to be done according to the theory.

However, the initial resudual can be obtained directly without first generating the matrix KEFT. So, I made this change to reduce memory usage.

Because of the elimination of KEFT, now I had to move the generation of the vector of prescribed voltages, VEF, to before calculating the initial residual. Because of this, I changed the function that generated VEF so that it does more processing than before and the function that calls the generation of VEF is not too long.