Input file reding
I modified the input file reading so that it allows to not have a section if it is not needed (e.g., CNT geometry when using only GNPs). I actually used the while loop of the old version.

I also added the initialization of the resistances flags in simulation parameters, which I had forgotten to do.