Changed resistance by its invers in functions that fill 2D sparse matrix
The functions that add elements to the 2D sparse stiffness matrix calcualted the inverse of resistance. This because the inverse of resistance is the value that needs to be added to the stiffness matrix.

However, I noticed that I was always calculating the inverse even when using unit resistors. In such case, calculating the invers is not necessary. So I modified the functions that add elements to the stiffness matrix to take the invers of resistance as input parameter directly. The inverse of resistance is calculated before calling this function and only when the resistor flag is set to 1. This saves a bit of computational time and likely reduces numerical errors a little bit. However, these improvements might not be noticeable.