Search direction vector P calculated in CG
The search direction vector P that is used in the conjugate gradient (CG) was calculated while generating the vectors for the sparse stiffness matrix. This made sense before adding the preconditioner.

After adding a preconditioner, the search direction is actually calculated during CG. So, the previous calculation of P is overriden. Thus, it is useless. So I removed the calculation of P from the function that generates the vectors for the sparse stifness matrix.