Correction in simplices in touch
Given the correction to HK76 when storing the vertices of simplices, the same correction was needed in the Generate_Network class where something similar is done. In the Generate_Network class the simplices that are touching are found. The vertices of the simplices were saved in vectors. However, as in the previous update, this will result in repeated vertices. Thus, now I am using sets.