Corrected window extration for GNPs and exporting clusters and percolated clusters
Some corrections were made in the Cutoff_Wins class

1) I changed the way in which the functions Find_intersections_of_gnp_edges_with_window_boundaries and Find_intersections_of_window_edges_with_gnp_faces iterated over edges. Before I created an adjacency matrix and a visited matrix of the same size. However I ralized that this matrix cannot be symmertic beacause it is not squared. I should have a segmentation fault error or something but I was getting none. So in both cases I created an array of the Edge data straucture to iterate over all edges of either a window or GNP.

2) In the function Does_edge_intersect_gnp I noticed I was using integers instead of Point_3D in some opeartions. Also I was not using the correct vertex: sometimes I was using a GNP vertex instead of a window vertex and vice versa.

Also, the clusters of GNPs were not being exported because I was using class variables that actually were empty. SI I modified the function that exported clusters to take as input the variable of clusters. In this way now I can use the temporary vectors that have the clusters before determining percolation.

Added some commnets for debugging