Corrected a comparison with zero
In comparing a quantity to zero I did not considered the case of negative numbers. The issue here is that I needed the quantity to be close to zero not just less than zero. To achieve this I need to compare the absolute value with zero. In this way a negative number does not make true the if-statement.