Corrected boundary points in non-prescribed boundaries that were ignored
All GNP boundary points were ignored when assigning a node number to a GNP point. However, this is not useful unless the six boundaries have prescribed voltages. Otherwise, some points will not have a node assigned. Thus, I changed the if-statement that caused all boundary points to be ignored. Now when a GNP point is assigned a node number it is checked if it already has a reserved node number assigned.

Because of this change, the variable that contains the number of GNP points at the boundary (n_gnp_pts_b) is useless now. So I deleted it.

Also, I noticed that I was using a variable with name "GNP" but I changed it to "gnp_i" because GNP is a data structure now.