Small update: casting and constructur in Electrical_analysis
I realized there was an unnecessary casting in the function Currents_through_boundary_gnps in the class Electrical_analysis. A casting from int to long int was made because the integer variable was an input to the function. When that function was called, I was initially using a variable. That changed to having either 0 or 1. Thus, I changed the variable type to long int so that the casting is unnecessary.

I noticed that the class Electrical_analysis did not have a constructor. Thus, I added its constructor.

I also added some comments for debugging.