Made some changes in Generate_Network
Made some changes in the class Generate_Network to condense some lines into functions and reduce repeated code.

I also made some adjustements on how I was using some vectors of known size. Instead of making them grow, I created them with the size I needed.

In the functions for one and two overlapping points, I added some lines to deal with the case when a point is the same as another point.