Visualization files and tolerance
Visualization files now are grouped in folders depending in what type of visualization they are. There is one folder per observation window.

I also modified a bit the filenaming to make names shorter.

I added the tolerance for the conjugate gradient algorithm as an input parameter so that it can be modified if needed. In some systems it may be useful to do so.