Typo in section Visualization_Flags
The section of the input file for the visualization flags had as name Visualization_flags. This is inconsistent with the rest of the sections in the input file, where every word is capitalized. Thus, I changed the name of the section to Visualization_Flags.

Also made some changes in the input file to correct typos or rewrite sentences.