Added the option for CNT_deposit
Added the option of CNT_deposit in the input file, with some comments on some options that are ignored. Also, when the input file is reaed, this option is considered. Also, the geometry of the extended domain is different for CNT_deposit.

The way the observation window is updated is different when CNT_deposit is chosen. In such case, the window geometry along the z-direction does not chage.

For the classes Contact_grid, Cutoff_Wins, and Hoshen_Kopelman I just updated the case when there are GNPs. This case used to be when the particle type was not CNT_wires. Now this case happens when particle type is not CNT_wires nor CNT_deposit.

In the class Generate_Network I changed a hardcoded penetration flag to the penetration flag from the simulation parameters. This was acompanied by a note in the input file that the flag is ignored when CNT_deposit is chosen.