Avoid moving non-existing visualization files
If there is no percolation, visualization files for percolated clusters, backbone and dead branches will not be generated. However, the commands that move such files to the corresponding folders are still executed. This results in error messages from the command mv saying "No such file or directory".

To avoid having such error message, I changed the code to move the visualization files for percolated clusters, backbone, and dead branches only if there are elements in the clusters_cnt or clusters_gnp vectors. As a consequence, if there are only isolated particles, cluster files will not be generated and the command to move such non-existing files will not be esecuted.