Sample geometry and new direction in 2D
The function in Generate_Network that is used to assign global coordinates only works in the sample domain. So, if I need the global coordinates of points in the boundary layer, then I would need to duplicate the functions. In these duplicated functions the extended domain would be used instead of the sample domain.

To avoid repeating code, I instead generate a new Geom_sample object that contains the sample domain as the extended domain. In this way I avoid modifying any function. I also removed the usage of some unused variables in the Geom_sample data structure that also contain the sample geometry.

I also updated the z-coordinate of the extended domain's "origin" and its height (length along z) for the case of CNT deposit. For a CNT deposit, these quantities have to be calculated differently.

The angle for the new direction in 2D was always positive due to an absolute value operation that was carried over from the 3D version of this function. I just removed the absolute value while calculating the new angle to also obtain negative angles.