Removed variables form Geom_sample data structure
In the Geom_sample data structure there is a cuboid with the sample geometry. However, there are also variables that contain the sample geometry. I deleted these variables as the information is redundant. I was still using these repeated variables in other places in the code so I had to change that. Not the cuboid for the sample geometry is always used.