Sample geometry for deposit and unused functions
I had forgotted to actually use the new Geom_sample data structure when generating a CNT deposit.

I am also adding some function that I was intending to use for the non-penetration of CNTs in the deposit. However, I developed a new method that makes more sense to me. So I am adding this functios just for the record. In the following update these functions will be deleted.