Torus approach (using angles)
In this update I used a torus to define all possible positions of a new point. Then I calculated the distance of a point to this torus by minimizing the distances, which was a function of two angles. Then I would get the angles that result in minimum distance. However this approach did not work.